BrainRot/backend/requirements.txt
fastapi
uvicorn[standard]
pillow
gtts
requests
python-multipart
openai
